Preserve scaling parameters even if scaling is disabled in decodeImageFrame
Prescaling of images gets disabled in decodeImageFrame if
option.preScale.enabled is false, or if the device can't render floats
(eg on iOS devices). When this happens, preScale.scaled will be set to
false on the decoded frame. Accessing the scaling values is still useful
even if scaling wasn't done, so that viewers can adjust the VOI ranges
in consequence.